remove additional framework parameters from MPI command
This is a follow-up to the workaround in the previous commit.
The MPI command was failing with Python SDK tests due to the
string not being wrapped in quotes when part of the MPI command.
This commit is meant as a workaround until a better solution is
found for constructing the MPI command.